Refactor QuantumGraphExecutor class
